RELEASE 2.0.0 (#5)
* done: Testing ci/cd in heroku

* done: Testing ci/cd in heroku - enabled automatic deploy on branch

* done: Testing ci/cd in heroku - enabled automatic deploy on branch

* done: Testing ci/cd in heroku2

* wip: Testing the Continues Integration for the github

* wip: Testing if the node is installed

* wip: Testing if the location is int the root of the ff project

* wip: Using the npm build

* wip: Use the ignore file in standardjs

* wip: Edit workflow run-test.yml

* Create Testing Workflow

* wip: Add the Testing Laravel Framework

* wip: Adding comment in workflow laravel-test

* wip: Editing PHPInsights rule set

* done: Init insights

* done: Remove un used comment branch-address-form component

* done: Fix the admin-branches super admin bugfix

* done: Remove lodash module

* done: Create CustomerPackage Repository and fixing salon-table component

* done: Creation of BranchRepository + UserRoleRepository additional methods and Edit Branch AddressForm

* done: Initial Implementation for the salon table

* done: Fix psr12 standard

* done: Coding Standard

* done: Coding Standard

* bug: Fix the Admin/Packages issue on selection

* bug: Rename salon to management

* done: Coding Standard

* done: Fix the management table pending the customer visitation reload event emission

* done: Fix the management table pending the customer visitation reload event emission

* done: Modify README.MD file logo center

* done: Fix the customer visitation error

* done: Fix app/Exception/Handler.php coding standard

* done: Add the icon for customer renew

* done: Fixing coding standard

* done: Implement the gym package new customer functionality

* done: Remove the .DS_Store file include also in global config && Implement the correct management table for gym

* done: Implement gym management table design + js functionality + GymVisitationType Enums and DB migration for gym_visitation

* done: Implement the GymVisitation migration + model + apiController

* done: Adding the gymvisitation store method

* done: Connect the front end interaction from the backend gymvisitation

* done: Fix the coding standard for gym visitation

* done: Fix the issue on superadmin  change email module

* done: Fix the issue on GYM IN/OUT also add management branch name and type for reference

* done: Update README.MD include the command for running the insights

* done: Implement the default main branch type salon

* done: Implement the default branch in Main Branch but exclude it to the Nav Branch Counter

* done: Implementation of sadmin in .env

* fix: Fix the Admin Branches issue for the super admin branch

* fix: Un used DB facades in BranchRepository

* fix: Fix the improper customer visitation on freebies

* done: Added new Payment Type Cash

* done: Implementation of GymVisitation@store api endpoint

* done: Implementation of Management table gym IN/OUT status

* done: Fix the today visitation status in gym visitation

* done: Fix grammar error in gymvisitation

* done: init the customer package visitation

* 1. Edit Branch Name default F&F.Co Main Office.
2. Add Payment Type CASH.
3. Edit the log file in checkCustomerPackageExpiration.

* 1. Include PaymentType cash
2. Add Cash Payment Type

* Change the main branch name to the F&F.Co Main Office

* done: Fix the error in same ref number (Add New Customer, Customer Renew)

* done: Filter the Package type in Search Management Table && add a old login.blade old field input type

* 1. Rename home to management-home in view.
2. Added the cookie for old_hometype to determine the old Account Type selection by the user

* 1. Implement the mid night auto logout for the gym services

* Implement the midnight auto logout for the gym service

* Initial Implementation of revised customer package query

* Add package_type field in customer_package table

* Remove unnecessary imports  in create_customer_package Carbon/Carbon

* Lint psr12

* Add the customer_package.package_type to determine easily what is the main package_type of a customer_package edit the following files: app/CustomerPackage, app/Http/Controllers/Api/CustomerPackageController, app/Http/Requests/CustomerPackage

* 1. Fix the Request\CustomerPackage package_type field.
2. Fix the SaveNewCustomerJS include package_type in customerPackage Request

* 1. Edit the CustomerPackageRepository::getAll
2. filter using customer_package_id desc (latest first)
3. filter using the CustomerPackage.package_type

* 1. Change gym visitation error for customer not yet visited
2. Cleanup the ManagementTable livewire component only => mount&&onSearchText
3. Remove unnecessary condition query on CustomerRepository::getAll method
4. Create management-table-row laravel component
5. Retain only new customer + renew customer for ManagementAction component

* 1. Fix the management-table-row component price line and package name position
2. Fix the CustomerVisits using the gym package type

* 1. Implement the ManagementTable component onSearch (Reference NO || Customer Name)
2. Implement the loading for ManagementTable

* 1. Rename Loading data -> Loading records
2. Creation of customer-list livewire component

* 1. Implement the customer-list component search
2. Create the customerPackage Repository search customerPackage Method

* Enable the service only when it is active

* Add the tag for expired package

* Edit the customer-list component renew button

* Implement the new Package (Front-End Validation)

* Implement Admin Initial Responsive

* Implement responsive login

* Implement Nav && Management-Table responsive

* Remove Admin Search useless code

* Responsive: Login form.

* Responsive: Login form.

* Responsive: Login form. (Mobile + Tablet + Phone)

* Responsive: Nav component (Mobile, Tablet, Desktop)

* Define variables for sm-xl breakpoints

* Fixed the Responsiveness Management Area

* admin.packages d-none

* Replace common js require to import utils = ChangeUserEmail

* Replace common js require to import utils all used js files

* Implement the d-none user-profile in nav component

* Implement the 850 max-height media to add 5rem padding bottom spacing

* Remove the No of visits when the Package Type is Gym

* Fix Package Request, Include validation for package_type in PackageController API

* Increase the migrattion size for create_package

* Add exit button in renew page

* Implement the automatic expired on customer package  whose visitation is 0 to zero

* Implement the automatic redirection on table after saving the new customer

* Fix the improper table-management-row

* Fix the improper management table after the adding new customer

* Change the save to ADD VISITS/ ADD CUSTOMER

* Bugfix the improper (gym package)customer data

* Re enable the public/app.js && public/mix-manifest.json

+++ database/migrations/2020_07_12_065427_create_package_table.php
@@ -19,11 +19,11 @@ class CreatePackageTable extends Migration
             $table->enum('package_status', PackageStatus::getValues())->default(PackageStatus::INACTIVE);
             $table->enum('package_type', PackageType::getValues());
             $table->tinyInteger('salon_no_of_visits')->unsigned(); // visitation (number) for management.
-            $table->tinyInteger('salon_days_valid_count')->unsigned(); // number of days that the valid will be valid
+            $table->mediumInteger('salon_days_valid_count')->unsigned(); // number of days that the valid will be valid
             $table->tinyInteger('gym_no_of_visits')->unsigned()->default(0); // if 0 then it is only days
-            $table->tinyInteger('gym_days_valid_count')->unsigned();
+            $table->mediumInteger('gym_days_valid_count')->unsigned();
             $table->tinyInteger('spa_no_of_visits')->unsigned();
-            $table->tinyInteger('spa_days_valid_count');
+            $table->mediumInteger('spa_days_valid_count');
             $table->timestamps();
         });
